Add Contact Class
- To achieve learning objective [Inheritance] 
- Extracted commonalities from Address, Phone and Email classes 
- Modified Address class to extend Contact Class
